Add Tenant REST functionality

+++ employee-rostering-backend/src/main/java/org/optaweb/employeerostering/service/tenant/TenantRepository.java
@@ -24,7 +24,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TenantRepository extends JpaRepository<Tenant, Long> {
+public interface TenantRepository extends JpaRepository<Tenant, Integer> {
 
     @Query("select t from Tenant t " +
             // Deliberately order by id instead of name to use generated order
